% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_corrections_03.R
\name{amp_drift_corr}
\alias{amp_drift_corr}
\title{Charge Amplifier Drift Correction}
\usage{
amp_drift_corr(
  folder,
  tau = 9400,
  print.to.screen = FALSE,
  print.to.pdf = TRUE,
  res.reduction = 10,
  start.file.number = 1
)
}
\arguments{
\item{folder}{Path to folder containing measurements.}

\item{tau}{Numeric time constant of charge amplifier in the same time unit as the measurement data.
Default: \code{9400}}

\item{print.to.screen}{A logical value indicating if results should be plotted in the current R plot.
device. Slows down process. Default: \code{FALSE}.}

\item{print.to.pdf}{A logical value indicating  if results should be saved as PDFs. Does not slow
down the process as much as printing to the R plot device and is considered necessary to quality check
the results. Default: \code{TRUE}.}

\item{res.reduction}{A numeric value to reduce the number of time steps by during plotting. Speeds up
the plotting process and reduces PDF size. Has no effect on the results, only on the plots. Default: \code{10}.}

\item{start.file.number}{A numeric value indicating at which file the loop should start. Helpful in
case the loop stopped during a previous run. Default: \code{1}.}
}
\value{
The following folders will be created (if they did not exist before):
\itemize{
\item \code{"./ampdriftcorr/"}
\item \code{"./ampdriftcorr/logs/"}
\item \code{"./ampdriftcorr/pdfs/"}
}

Saves amplifier-drift-corrected data files in CSV-format in new subfolder \code{"./ampdriftcorr"},
log files containing information on the method used to correct the amplifier drift in
\code{"./ampdriftcorr/logs/"} and PDFs with the graphs before and after correction in \code{"./ampdriftcorr/pdfs/"}.
}
\description{
Removes the systemic, asymptotical drift of charge amplifiers with resistor-capacitor (RC) circuits.
}
\details{
The input files need to be in the following format:\tabular{ccc}{
   \code{t} \tab  \tab \code{y} \cr
   \code{t.1} \tab  \tab \code{y.2} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}
}
\examples{
\dontrun{
# Using package example files from GitHub stored within
# https://github.com/Peter-T-Ruehr/forceR-data/blob/main/example_data.zip

# define folder where files are are stored that need amplifier drift correction
data.folder <- "./example_data"
cropped.folder <- file.path(data.folder, "cropped")

# run apmplifier drift correction on all of files in cropped.folder
# the results will be saved in several new folders within the cropped folder.
# If print.to.pdf == TRUE, than graphs showing the corrections will be saved as PDFs as well.

amp_drift_corr(folder = cropped.folder,
               tau = 9400,
               print.to.screen = FALSE,
               print.to.pdf = TRUE,
               res.reduction = 10,
               start.file.number = 1)
}
}
