% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{avg_peaks}
\alias{avg_peaks}
\title{Average Curves per Group}
\usage{
avg_peaks(df)
}
\arguments{
\item{df}{The resulting tibble of the function \code{red_peaks_100()}. See \code{?red_peaks_100} for more details.}
}
\value{
This function returns a tibble made of three columns: \code{species} containing the species names, \code{index} ranging from 1 to 100 for each species,
and \code{force.norm.100} containing the averaged and rescaled curve of each species.
}
\description{
Calculates mean curve shape per group (here: species) and rescales result on the y axis to range from 0 to 1.
}
\examples{
# This example contains a self-sufficient data PREPARATION section
# before the function is actually run.
require(dplyr)

# PREPARATION ####
# create a classifier to store specimen info (see package vignette for details)
classifier <- tibble(species = c("A","A","A","A","B","B","B","B"),
                     specimen = c("a","a","b","b","c","c","d","d"),
                     measurement = paste0("m_0", 1:8),
                     amp = c(rep(2,4), rep(0.5, 4)),
                     lever.ratio = rep(0.5, 8))

# create temporary tibble to store data for bite series simulation
classifier.temp <- classifier \%>\%
  mutate(type = c(rep("sin", 4), rep("plat", 4)),
         max.y = c(1.9, 2.4, 2.2, 2.0, 6.8, 7.2, 7.5, 7.2),
         length.of.bite = c(20, 20, 18, 22, 50, 40, 45, 40),
         length.of.series = c(rep(200, 4), rep(600, 4)),
         jit = c(rep(0.5, 4), rep(1, 4)))

# create tibble with simulated time series with different
# bite characteristics for each measurement, specimen and species
df.all <- NULL
for(i in 1:nrow(classifier.temp)){
  df.curr <- simulate_bites(no.of.bites = 7,
                            length.of.bite = classifier.temp$length.of.bite[i],
                            length.of.series = classifier.temp$length.of.series[i],
                            max.y = classifier.temp$max.y[i],
                            max.y.jit = 15,
                            jit = classifier.temp$jit[i],
                            bite.type = classifier.temp$type[i],
                            plot = FALSE)

  # add measurement number to df.curr
  df.curr <- df.curr \%>\%
    mutate(measurement = classifier.temp$measurement[i])

  # add current sumulated bite series to df.all
  df.all <- rbind(df.all, df.curr)
}
# remove temporary tibble to avoid confusion
rm(classifier.temp)

# rename columns
df.all <- df.all \%>\%
  rename(force = y)

# add classifier info to bite table (df.all)
df.all <- left_join(df.all,
                    classifier,
                    by = "measurement")

peaks.df <- find_strongest_peaks(df = df.all,
                                 no.of.peaks = 5,
                                 print.to.pdf = FALSE)

# rescale bites
peaks.df.norm <- rescale_peaks(df.peaks = peaks.df,
                                df.data = df.all)

# reduce to 100 obesrvations per bite
peaks.df.norm.100 <- red_peaks_100(df = peaks.df.norm,
                                   path.plots = path.plots,
                                   print.to.pdf = FALSE)

# RUN THE FUNCTION ####
peaks.df.100.avg <- avg_peaks(df = peaks.df.norm.100)

# plot results
par(mfrow = c(1,2))
plot(peaks.df.norm$t.norm,
     peaks.df.norm$force.norm,
     type="l")

plot(peaks.df.100.avg$index,
     peaks.df.100.avg$force.norm.100,
     type="l")
par(mfrow = c(1,1))
}
