% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_corrections_03.R
\name{baseline_corr}
\alias{baseline_corr}
\title{Automatic or Manual Baseline Correction of Time Series}
\usage{
baseline_corr(
  file,
  corr.type = "auto",
  print.to.screen = TRUE,
  print.to.pdf = TRUE,
  window.size.mins = 1000,
  window.size.means = NULL,
  quantile.size = 0.05,
  y.scale = 0.5,
  res.reduction = 10,
  Hz = 100
)
}
\arguments{
\item{file}{A character string containing the full path to the measurement file that needs correction.
See Details for info on what the file should look like.}

\item{corr.type}{Character string defining the desired mode of baseline correction. One of \code{"auto"}
or \code{"manual"}. Default: \code{"auto"}}

\item{print.to.screen}{A logical value indicating if results should be plotted in the current R plot
device. Default: \code{TRUE}.}

\item{print.to.pdf}{A logical value indicating  if results should be saved as PDFs. This is considered
necessary to quality check and communicate the results. Default: \code{TRUE}.}

\item{window.size.mins}{A numeric value for the size of the search window to find minima in. Should be
in the same time unit as the measurement. Longer peaks require higher values, shorter peaks require
smaller values. Default: \code{1000}.}

\item{window.size.means}{A numeric value for the size of the window to average the minima in. Should be
in the same time unit as the measurement. By default (\code{NULL}), the same value as specified for
\code{window.size.mins} is used.}

\item{quantile.size}{A numerical value between \code{0} and \code{1} to define the quantile which is treated as
the 'minimum' of a sliding window. Default: \code{0.05}.}

\item{y.scale}{A numeric value to reduce the y-axis range during plotting. This simplifies the manual
placement of the points during the manual correction procedure.}

\item{res.reduction}{A numeric value to reduce the number of time steps by during plotting. Speeds up
the plotting process and reduces PDF size. Has no effect on the results, only on the plots. Default: 10.}

\item{Hz}{A numeric value to reduce sampling frequency for temporary analyses. This works as a smoothing
filter during temporary analyses and does not reduce the actual sampling frequency of the data.
Default: \code{100}.}
}
\value{
The following folders will be created (if they did not exist before):
\itemize{
\item \code{"./ampdriftcorr/"}
\item \code{"./ampdriftcorr/logs/"}
\item \code{"./ampdriftcorr/pdfs/"}
}

Saves amplifier-drift-corrected data files in CSV-format in new subfolder \code{"./ampdriftcorr"},
log files containing information on the method used to correct the amplifier drift in
\code{"./ampdriftcorr/logs/"} and PDFs with the graphs before and after correction in \code{"./ampdriftcorr/pdfs/"}.
ampdriftcorr.folder <- file.path(data.folder, "cropped/ampdriftcorr")
}
\description{
If baseline (zero-line) of measurement is unstable (e.g. due to temperature fluctuations, wind, ...),
the baseline needs to be continually adjusted throughout the measurement. This script allows an
automatic adjustment of the baseline.
The automatic approach invokes a sliding window, during which the 'minimum' within each sliding window
is stored. A 'minimum' is defined by the \code{quantile.size}: if set to \code{0.05}, the value below which only 5\% of
the measurement data within the sliding window lies, is treated as the current window's minimum. This
prevents the treatment of potential artifacts as minima. In a second iteration, another sliding window
calculates the average of these 'minima'. The resulting values are subtracted from
the original time series. This approach works well for time series with relatively short peaks.
If the automatic approach does not yield acceptable results, an interactive manual approach to correct
the baseline can be performed instead.
}
\details{
The input files should to be in the following format:\tabular{ccc}{
   \code{t} \tab  \tab \code{y} \cr
   \code{t.1} \tab  \tab \code{y.2} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}


In case there are more than two columns, only the first two columns will be used. If the first two columns
are not named 't' and 'y', they will be renamed.
}
\examples{
\dontrun{
# Using a package example file from GitHub stored within
# https://github.com/Peter-T-Ruehr/forceR-data/blob/main/example_data.zip

#'########### AUTOMATIC MODE

# define folder and file to apply the baseline drift correction to
data.folder <- "./example_data"
ampdriftcorr.folder <- file.path(data.folder, "cropped/ampdriftcorr")
file = file.path(ampdriftcorr.folder, "1068_cropped_ampdriftcorr.csv")

# plot file
plot_measurement(file)

# run the automatic baseline drift correction
# the results will be saved in new subfolders of the ampdriftcorr.folder
baseline_corr(file,
             corr.type = "auto",
             print.to.screen = TRUE,
             print.to.pdf = TRUE,
             window.size.mins = 2000,
             window.size.means = NULL,
             quantile.size = 0.05,
             y.scale = 0.5,
             res.reduction = 10,
             Hz = 100)

[1] "1068_cropped_ampdriftcorr"
[1] "You selected automatic drift correction:"
[1] "sliding minima window size: 2000"
[1] "sliding means of minima window size: 2000"
[1] "calculating with 100 Hz."
[1] "Finding 5th percentile sliding minima..."
100\%...[1] "Finding sliding means..."



#'########### MANUAL MODE

#'# define folder and file to apply the baseline drift correction to
data.folder <- "./example_data"
ampdriftcorr.folder <- file.path(data.folder, "cropped/ampdriftcorr")
file = file.path(ampdriftcorr.folder, "1174_cropped_ampdriftcorr")

# plot file
plot_measurement(file)

# run the manual baseline drift correction.The user is expected to
# interactively mark the baseline with dots in the R plot device.
# the results will be saved in new subfolders of the ampdriftcorr.folder

baseline_corr(file,
             corr.type = "manual",
             print.to.screen = TRUE,
             print.to.pdf = TRUE)
             plot_measurement(file)

[1] "1174_cropped_ampdriftcorr"
[1] "You selected manual drift correction."
[1] "Please select baseline points and click \"Finish\"."
[1] "31 line points taken."
[1] "Done!"
}
}
