% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropping_02.R
\name{convert_measurement}
\alias{convert_measurement}
\title{Convert LJStream *.dat file to standard time series}
\usage{
convert_measurement(file, decimal = ".")
}
\arguments{
\item{file}{File path to raw measurement (*.dat file).}

\item{decimal}{A character string of length 1 defining the decimal separator. Default \code{"."}.}
}
\value{
Saves data file in CSV-format in new subfolder "./converted". If this folder does not exist, it will be created. \cr
\cr
The output file has the following format:\tabular{ccc}{
   \strong{\code{t}} \tab  \tab \strong{\code{y}} \cr
   \code{t.1} \tab  \tab \code{y.1} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}
}
\description{
Converts LJStream *.dat file to standard time series.
}
