% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropping_02.R
\name{crop_measurement}
\alias{crop_measurement}
\title{Crop Time Series}
\usage{
crop_measurement(file, decimal = ".")
}
\arguments{
\item{file}{File path to measurement.}

\item{decimal}{A character string of length 1 defining the decimal separator. Default \code{"."}.}
}
\value{
Saves cropped data file in CSV-format in new subfolder "./cropped". If this folder does not exist, it will be created. \cr
\cr
The output file has the following format:\tabular{ccc}{
   \strong{\code{t}} \tab  \tab \strong{\code{y}} \cr
   \code{t.1} \tab  \tab \code{y.1} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}
}
\description{
Interactive function to crop a time series.
}
\details{
Select points at start and end of desired part of measurements. Only the last two points will be
taken into account to allow the user to correct erroneous clicks.
}
\examples{
\dontrun{
# Using a package example file from GitHub stored within
# https://github.com/Peter-T-Ruehr/forceR-data/blob/main/example_data.zip

# load the original file
data.folder <- "./example_data"
file <- file.path(data.folder, "0982.csv")

# crop file
crop_measurement(file)

# define folder where cropped file is stored
cropped.folder <- file.path(data.folder, "cropped")

# plot the cropped file
cropped.file <- file.path(cropped.folder, "0982_cropped.csv")
plot_measurement(cropped.file)
}
}
