% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropping_02.R
\name{crop_measurement_LJStream}
\alias{crop_measurement_LJStream}
\title{Crops Time Series}
\usage{
crop_measurement_LJStream(file, decimal = ".")
}
\arguments{
\item{file}{File path to measurement.}

\item{decimal}{A character string of length 1 defining the decimal separator. Default \code{"."}.}
}
\value{
Saves cropped data file in CSV-format in new subfolder "./cropped". If this folder does not exist, it will be created. \cr
\cr
The output file has the following format:\tabular{ccc}{
   \strong{\code{t}} \tab  \tab \strong{\code{y}} \cr
   \code{t.1} \tab  \tab \code{y.1} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}
}
\description{
Interactive function to crop a time series measured with LJStream (LabJack Corporation, Lakewood, Colorado, US), the standard software of the forceR setup.
}
\details{
Select points at start and end of desired part of measurements. Only the last two points will be
taken into account to allow the user to correct erroneous clicks.
}
