% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.all.200}
\alias{df.all.200}
\title{Simulated Time Series - e.g. Bite Force Measurements}
\format{
A data frame with 1,944 rows and 3 columns:
\describe{
\item{t}{time, in ms}
\item{y}{measured values, in V}
\item{measurement}{measurement names}
}
}
\usage{
df.all.200
}
\description{
A series of 24 measurements with six simulated peaks each.
Simulated using the \code{simulate_bites()} function. Reduced to a sampling
frequency of 200 Hz with \code{reduce_frq()}. Fits the \code{classifier} sheet that
is also part of the \code{forceR} package.
}
\keyword{datasets}
