% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.all.200.tax}
\alias{df.all.200.tax}
\title{Simulated Force Measurements with Taxonomic Info.}
\format{
A data frame with 1,944 rows and 5 columns:
\describe{
\item{species}{species names}
\item{specimen}{specimen names}
\item{measurement}{measurement names}
\item{t}{time, in ms}
\item{force}{measured values, in N}
}
}
\usage{
df.all.200.tax
}
\description{
A series of 24 measurements with six simulated peaks each.
Simulated using the \code{simulate_bites()} function and translated into force
values with \code{y_to_force()}. Fits the \code{classifier} sheet that
is also part of the \code{forceR} package.
}
\keyword{datasets}
