% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.all}
\alias{df.all}
\title{Simulated Force Measurements with Taxonomic Info.}
\format{
A data frame with 9,600 rows and 3 columns:
\describe{
\item{t}{time, in ms}
\item{y}{measured values, in V}
\item{measurement}{measurement names}
}
}
\usage{
df.all
}
\description{
A series of 24 measurements with six simulated peaks each.
Simulated using the \code{simulate_bites()} function .
Fits the \code{classifier} sheet that is also part of the \code{forceR} package.
}
\keyword{datasets}
