% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcer.R
\name{forcer-shiny}
\alias{forcer-shiny}
\alias{forcerOutput}
\alias{renderForcer}
\alias{forcer_html}
\title{Shiny bindings for forcer}
\usage{
forcerOutput(outputId, width = "100\%\%", height = "400px")

renderForcer(expr, env = parent.frame(), quoted = FALSE)

forcer_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a forcer}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using forcer within Shiny
applications and interactive Rmd documents.
}
