% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcer.R
\name{forcer}
\alias{forcer}
\title{Create Force Directed Graph}
\usage{
forcer(data = NULL, forcetype = c("ReactForce2D", "ReactForce3D",
  "ReactForceVR"), ..., width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{\code{json} format as described in \url{https://github.com/vasturiano/react-force-graph#input-json-syntax}.
Use \code{d3r::d3_igraph} to convert \code{igraph} networks into the proper format.}

\item{forcetype}{\code{string} one of \code{"ForceGraph2D"}, \code{"ForceGraph3D"}, or \code{"ForceGraphVR"}}

\item{...}{other props for the force graph; see \url{https://github.com/vasturiano/react-force-graph#api-reference}}

\item{width, height}{\code{string} valid \code{CSS} size unit}

\item{elementId}{\code{string} valid \code{CSS} selector id}
}
\description{
Create Force Directed Graph
}
\examples{
library(d3r)
library(igraph)
library(forcer)

# VR crashed my RStudio so open in real browser
options(viewer = NULL)

# 2D
forcer(
  data = d3r::d3_igraph(igraph::make_tree(60,3,mode="in")),
  forcetype = "ForceGraph2D",
  nodeColor = list("#0f0"),
  linkCurvature = 0.25
)

# 3D
forcer(
  data = d3r::d3_igraph(igraph::make_tree(60,3)),
  forcetype = "ForceGraph3D"
)

# VR
forcer(
  data = d3r::d3_igraph(igraph::make_tree(60,3)),
  forcetype = "ForceGraphVR"
)
}
