% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{forceR_example}
\alias{forceR_example}
\title{Get path to forceR example}
\usage{
forceR_example(type = "folder")
}
\arguments{
\item{type}{A character string (either \code{"folder"}, \code{"raw"}, or
\code{"ampdriftcorr"}) defining if the path returned be the function
should point to one of the the example files or the folder containing them.
Default: \code{"folder"}.}
}
\value{
If \code{type = "folder"}: returns the file path to the folder containing
\code{BF_raw.csv} and
\code{BF_ampdriftcorr.csv}.

If \code{type = "LJStream"}: returns the file path
to \code{BF_raw.csv}, which contains a short bite force raw
measurement.

If \code{type = "raw"}: returns the file path
to \code{BF_raw.csv}, which contains a short bite force raw
measurement.

If \code{type = "ampdriftcorr"}: returns the file path
to \code{BF_ampdriftcorr.csv}, which contains a short bite
force raw measurement where the amplifier drift as been corrected for by
the amp_drift_corr() funcion.
}
\description{
forceR comes with example files of short bite force measurements.
The files are stored in \verb{forceR's} \code{inst/extdata}
folder, and this function returns the path to that folder or one of the
files  so they can be used in examples.
}
