% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models}
\alias{models}
\title{Polynomial Models Describing Average Peak Shapes.}
\format{
A list with 4 named entries, each containing the 4-degree polynomial
model that describes the average peak shape of the given species.
}
\usage{
models
}
\description{
Extracted from \code{df.all}.
}
\keyword{datasets}
