% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks.df.100.avg}
\alias{peaks.df.100.avg}
\title{Average Peak Shapes per Species}
\format{
A data frame with 400 rows and 3 columns:
\describe{
\item{species}{species names}
\item{index}{values from 1:100 in each species}
\item{force.norm.100.avg}{normalized values reduced to 100 observations per
species, in N}
}
}
\usage{
peaks.df.100.avg
}
\description{
Normalized force values describing the average shape of the 5 strongest peaks
per species of \code{df.all}
}
\keyword{datasets}
