% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks.df.norm.100}
\alias{peaks.df.norm.100}
\title{Normalized Peak Shapes with 100 Observations}
\format{
A data frame with 2000 rows and 6 columns:
\describe{
\item{species}{species names}
\item{measurement}{measurement names}
\item{specimen}{specimen names}
\item{peak}{peak numbers}
\item{index}{values from 1:100 in each measurement}
\item{force.norm.100}{normalized values reduced to 100 observations per
measurement, in N}
}
}
\usage{
peaks.df.norm.100
}
\description{
Normalized force values describing the shapes of all 5 strongest peaks
per species of \code{df.all}, reduced to 100 observations per measurement.
}
\details{
Result of \code{rescale_peaks()}.
}
\keyword{datasets}
