% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks.df.norm}
\alias{peaks.df.norm}
\title{Normalized Peak Shapes}
\format{
A data frame with 223 rows and 6 columns:
\describe{
\item{measurement}{measurement names}
\item{peak}{peak numbers}
\item{t.norm}{time values from 0:1 in each measurement, in ms}
\item{force.norm}{force values from 0:1 in each measurement, in N}
\item{species}{species names}
\item{specimen}{specimen names}
}
}
\usage{
peaks.df.norm
}
\description{
Normalized force values describing the shapes of all 5 strongest peaks
per species of \code{df.all}
}
\details{
Result of \code{rescale_peaks()}.
}
\keyword{datasets}
