% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_measurement}
\alias{plot_measurement}
\title{Plot raw measurement}
\usage{
plot_measurement(file, columns = c(1:2))
}
\arguments{
\item{file}{File path to measurement.}

\item{columns}{A vector of column numbers. The first entry will be used as the x-axis values, the second entry as y-axis values.
All other columns will be ignored. Default: \code{c(1,2)}.}
}
\value{
Creates a plot in the current plot device.
}
\description{
Plots a time series.
}
\details{
#' The input files need to be in the following format (even though column names do not matter):\tabular{ccc}{
   \strong{\code{t}} \tab  \tab \strong{\code{y}} \cr
   \code{t.1} \tab  \tab \code{y.1} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}
}
\examples{
filename = forceR_example(type = "raw")
plot_measurement(filename)

}
