% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot Peaks}
\usage{
plot_peaks(
  df.peaks,
  df.data,
  additional.msecs = 2000,
  path.plots,
  print.to.pdf = TRUE
)
}
\arguments{
\item{df.peaks}{df.peaks The resulting tibble of the function \code{find_peaks()}. See \code{?find_peaks} for more details.}

\item{df.data}{A data frame or tibble in the below format. The columns \code{t} (time), \code{force} and \code{measurement}
(measurement ID) must be present. This will usually be the same table that was used before in \code{find_peaks()}.}

\item{additional.msecs}{A numeric value indicating how many m.secs before and after the actual peak curve should be plotted. Default: \code{2000}}

\item{path.plots}{A string character defining where to save the plots.}

\item{print.to.pdf}{A logical value indicating if the peak curves should be saved as PDFs. Default: \code{TRUE}
Default: \code{TRUE}}
}
\value{
In addition, if \code{print.to.pdf == TURE},
Plots one graph per peak curve and, if \code{print.to.pdf == TURE}, saves all peak curves as one PDF at \code{path.plots}.
}
\description{
Plots the peaks identified by the function \code{find_peaks()}.
}
\details{
#' The input data frame or tibble \code{df.data} needs to contain the following columns:\tabular{ccc}{
   \strong{\code{t}} \tab \strong{\code{force}} \tab \strong{\code{measurement}} \cr
   \code{t.1} \tab \code{force.1} \tab \code{measurement.1} \cr
   \code{...} \tab \code{...} \tab \code{...} \cr
   \code{t.n} \tab \code{force.n} \tab \code{measurement.m} \cr
}
}
\examples{
require(dplyr)

# PREPARATION ####
# create a classifier to store specimen info (see package vignette for details)
classifier <- tibble(species = c("A","A","A","A","B","B","B","B"),
                     specimen = c("a","a","b","b","c","c","d","d"),
                     measurement = paste0("m_0", 1:8),
                     amp = c(rep(2,4), rep(0.5, 4)),
                     lever.ratio = rep(0.5, 8))

# create temporary tibble to store data for bite series simulation
classifier.temp <- classifier \%>\%
  mutate(type = c(rep("sin", 4), rep("plat", 4)),
         max.y = c(1.9, 2.4, 2.2, 2.0, 6.8, 7.2, 7.5, 7.2),
         length.of.bite = c(20, 20, 18, 22, 50, 40, 45, 40),
         length.of.series = c(rep(200, 4), rep(600, 4)),
         jit = c(rep(0.5, 4), rep(1, 4)))

# create tibble with simulated time series with different
# bite charactersitics for each measurement, specimen and species
df.all <- NULL
for(i in 1:nrow(classifier.temp)){
  df.curr <- simulate_bites(no.of.bites = 7,
                            length.of.bite = classifier.temp$length.of.bite[i],
                            length.of.series = classifier.temp$length.of.series[i],
                            max.y = classifier.temp$max.y[i],
                            max.y.jit = 15,
                            jit = classifier.temp$jit[i],
                            bite.type = classifier.temp$type[i],
                            plot = FALSE)

  # add measurement number to df.curr
  df.curr <- df.curr \%>\%
    mutate(measurement = classifier.temp$measurement[i])

  # add current sumulated bite series to df.all
  df.all <- rbind(df.all, df.curr)
}
# remove temporary tibble to avoid confusion
rm(classifier.temp)

# rename columns
df.all <- df.all \%>\%
  rename(force = y)

# add classifier info to bite table (df.all)
df.all <- left_join(df.all,
                    classifier,
                    by = "measurement")

peaks.df <- find_strongest_peaks(df = df.all,
                                 no.of.peaks = 5,
                                 print.to.pdf = FALSE)

# RUN THE FUNCTION ####
plot_peaks(df.peaks = peaks.df,
           df.data = df.all,
           additional.msecs = 20,
           print.to.pdf = FALSE)
}
