% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{print_progress}
\alias{print_progress}
\title{Print progress}
\usage{
print_progress(current, end)
}
\arguments{
\item{current}{Numeric value of current loop iteration.}

\item{end}{Numeric value number last loop iteration.}
}
\value{
Prints percentage of loop progress to console.
}
\description{
Prints loop progress in \verb{[\%]} to console.
}
