% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{rescale_peaks}
\alias{rescale_peaks}
\title{Rescale Peaks}
\usage{
rescale_peaks(df.peaks, df.data, path.data = NULL)
}
\arguments{
\item{df.peaks}{The resulting tibble of the function \code{find_strongest_peaks()}. See \code{?find_strongest_peaks} for more details.}

\item{df.data}{A data frame or tibble in the below format. The columns \code{t} (time), \code{force} and \code{measurement}
(measurement ID) must be present. This will usually be the same table that was used before in \code{find_strongest_peaks()}.}

\item{path.data}{A string character defining where to save the results. If \code{NULL} (default),
data is not stored in file.}
}
\value{
This function returns a tibble with the same format as \code{df}, but with the additional columns \code{t.norm} and \code{force.norm} which will contain the
rescaled time and force data now both ranging from 0 to 1.
}
\description{
Rescales time series in x and y to values ranging from 0 to 1.
}
\section{\code{df} needs to contain the following columns:}{
\tabular{ccc}{
   \strong{\code{t}} \tab \strong{\code{force}} \tab \strong{\code{measurement}} \cr
   \code{t.1} \tab \code{force.1} \tab \code{measurement.1} \cr
   \code{...} \tab \code{...} \tab \code{...} \cr
   \code{t.n} \tab \code{force.n} \tab \code{measurement.m} \cr
}
}

\examples{
# Using the forceR::df.all.200.tax and forceR::df.all.200.tax datasets:

# rescale bites
peaks.df.norm <- rescale_peaks(df.peaks = peaks.df,
                                 df.data = df.all.200.tax)

# maximum values of time and force are both 1 in the whole data frame:
max(peaks.df.norm$t.norm)
max(peaks.df.norm$force.norm)

}
