% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rescale_to_range}
\alias{rescale_to_range}
\title{Scale data series to new minimum and maximum}
\usage{
rescale_to_range(data, from, to)
}
\arguments{
\item{data}{numeric vector containing the data to be scaled}

\item{from}{minimum of new range}

\item{to}{maximum of new range}
}
\value{
numeric vector with scaled data
}
\description{
Maps a series of numeric values to a new range defined by minimum (from) and maximum (to).
}
\examples{
rescale_to_range(data = 1:10,
                  from = 1,
                  to = 100)
}
