% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sorting.R
\name{sort_files}
\alias{sort_files}
\title{Sorts files after corrections}
\usage{
sort_files(data.folders, results.folder, move = FALSE)
}
\arguments{
\item{data.folders}{Character vector containing full folder paths of folders to check. This list must be sorted
according to the chronology of previous file editing. If a measurement exists in the last folder,
this is copied or moved into the \code{results.folder}, and files of the same measurement located in the
other folders will be ignored. Like this, the file of each measurement that underwent most correction steps
will be stored in the \code{results.folder}, while the rest of the files of the same measurement remain in place.}

\item{results.folder}{Character string defining the full path to the folder where the desired files will be stored.}

\item{move}{A logical value specifying if files should be moved (\code{move = TRUE}) or copied (\code{move = FALSE}). Default: \code{FALSE}}
}
\value{
This sorts existing files.
}
\description{
The files of each of the various possible correction steps (cropping, amplififer correction, drift correction)
are all located in their own folders. This function gets all files that represent the last correction step of a
given measurement out of all those folders and saves them in the \code{results.folder}.
}
\examples{
\dontrun{
# Using package example files from GitHub stored within
# https://github.com/Peter-T-Ruehr/forceR-data/blob/main/example_data.zip

# define data.folders
data.folders <- c(data.folder,
file.path(data.folder, "/cropped"),
file.path(data.folder, "/cropped/ampdriftcorr"),
file.path(data.folder, "/cropped/ampdriftcorr/baselinecorr"))

# define the folder to contain one corrected file per original raw measurement
results.folder <- file.path(data.folder, "/corrected/")

# run the file sorting
sort_files(data.folders = data.folders,
   results.folder = results.folder,
   move = FALSE)
}
}
