% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{summarize_measurements}
\alias{summarize_measurements}
\title{Summarize Table}
\usage{
summarize_measurements(df, var1, var2)
}
\arguments{
\item{df}{Data frame or tibble containing at least three colums. The column names must contain the
grouping variables defined in \code{var1} and \code{var2} and the column \code{force} (time series of force measurements).}

\item{var1}{A character string defining the column to calculate minimal and maximal force values per measurement.
This must be the column that contains the unique measurement ID, e.g. measurement number.}

\item{var2}{A character string defining the column for which the summary should be calculated.}
}
\value{
A tibble summarizing the input data frame \code{df}. The resulting tibble will contain
the columns \code{t}, \code{force}, \code{measurement}, \code{species}, \code{specimen}, \code{amp}, \code{lever.ratio},
\code{max.F.measurement}, \code{mean.F.specimen}, \code{max.F.specimen}, \code{sdv.max.F.specimen}, \code{n.measurements.in.specimen}.
}
\description{
Finds mininum, maximum and standard deviation of force per measurement and taxon and
creates summary tibble.
}
\examples{
require(dplyr)

# PREPARATION ####
# create a classifier to store specimen info
classifier <- tibble(species = c("A","A","A","A","B","B","B","B"),
                     specimen = c("a","a","b","b","c","c","d","d"),
                     measurement = paste0("m_0", 1:8),
                     amp = c(rep(2,4), rep(0.5, 4)),
                     lever.ratio = rep(0.5, 8))

# create temporary tibble to store data for bite series simulation
classifier.temp <- classifier \%>\%
  mutate(type = c(rep("sin", 4), rep("plat", 4)),
         max.y = c(3.0, 2.6, 2.2, 2.0, 5.0, 5.2, 8.0, 7.2),
         length.of.bite = c(20, 20, 18, 22, 50, 40, 45, 40),
         length.of.series = c(rep(200, 4), rep(600, 4)),
         jit = c(rep(0.5, 4), rep(1, 4)))

# create tibble with simulated time series with different
# bite charactersitics for each measurement, specimen and species
df.all <- NULL
for(i in 1:nrow(classifier.temp)){
  df.curr <- simulate_bites(no.of.bites = 7,
                            length.of.bite = classifier.temp$length.of.bite[i],
                            length.of.series = classifier.temp$length.of.series[i],
                            max.y = classifier.temp$max.y[i],
                            max.y.jit = 15,
                            jit = classifier.temp$jit[i],
                            bite.type = classifier.temp$type[i],
                            plot = FALSE)

  # add measurement number to df.curr
  df.curr <- df.curr \%>\%
    mutate(measurement = classifier.temp$measurement[i])

  # add current sumulated bite series to df.all
  df.all <- rbind(df.all, df.curr)
}
# remove temporary tibble to avoid confusion
rm(classifier.temp)

# rename columns
df.all <- df.all \%>\%
rename(force = y)

# add classifier info to bite table (df.all)
df.all <- left_join(df.all,
                    classifier,
                    by = "measurement")

# RUN THE FUNCTION ####
# sumarize by measurement and specimen
df.all.summary <- summarize_measurements(df = df.all,
                       var1 = "measurement",
                       var2 = "specimen")


#  PLOT RESULT ####
\dontrun{
ggplot2::ggplot(data = df.all.summary, mapping = aes(x=specimen,y=max.F.measurement)) +
  geom_jitter(aes(color='blue'),alpha=0.7) +
  geom_boxplot(fill="bisque",color="black",alpha=0.3) +
  # scale_y_log10() +
  labs(y="max(F)/specimen") +
  guides(color=FALSE) +
  theme_minimal()
}
}
