% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{today}
\alias{today}
\title{Get Today's Date as String}
\usage{
today()
}
\value{
\tabular{ll}{
    \tab  \cr
   \code{date.string} \tab a character string of today's date in the format "yyyy_mm_dd" \cr
}
}
\description{
Creates a character string containing today's date
in the format "yyyy_mm_dd" which can be used in file names.
}
