% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropping.R
\name{convert_measurement}
\alias{convert_measurement}
\title{Convert LJStream *.dat file to standard time series}
\usage{
convert_measurement(file, write.files = TRUE)
}
\arguments{
\item{file}{File path to raw measurement (*.dat file).}

\item{write.files}{A logical value indicating if results should be
written as file. Default: \code{TRUE}.}
}
\value{
Returns and, if \code{write.files == TRUE}, saves converted data in
CSV-format in new subfolder "./converted". If this folder does not exist,
it will be created. \cr
\cr
The output tibble has the following format:\tabular{ccc}{
   \strong{\code{t}} \tab  \tab \strong{\code{y}} \cr
   \code{t.1} \tab  \tab \code{y.1} \cr
   \code{...} \tab  \tab \code{...} \cr
   \code{t.n} \tab  \tab \code{y.n} \cr
}
}
\description{
Converts LJStream *.dat file to standard time series.
}
\examples{
# get file path of forceR example file
filename <- forceR_example(type = "LJStream")
file.converted <- convert_measurement (file = filename,
                       write.files = FALSE)
file.converted

}
