% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{find_best_fits}
\alias{find_best_fits}
\title{Find Best Polynomial Fits for Curves}
\usage{
find_best_fits(df, path.data = NULL, path.plots = getwd(), print.to.pdf = TRUE)
}
\arguments{
\item{df}{The resulting tibble of the function \code{avg_peaks()}. See \code{?avg_peaks} for more details.}

\item{path.data}{A string character defining where to save the results and log files. If \code{NULL} (default),
data is not stored in a file.}

\item{path.plots}{A string character defining where to save the plots. Default: working directory.}

\item{print.to.pdf}{A logical value indicating if the results of the initial peak finding should be saved as PDFs. Default: \code{TRUE}}
}
\value{
Returns the a numerical value representing the number of coefficient that was most often under the first 4 models that were followed by an
AIC-change \verb{<= 5\%} by the next model. Additionally, plots showing the model fits and a histogram of the coefficients that met the 5\%-criterion are either plotted to
the plot device (\code{print.to.pdf = FALSE}) or saved as PDFs in \code{path.plots} (\code{print.to.pdf = TRUE}).
}
\description{
Calculates best model fits for all curves based on AIC criterion. The function fits polynomial functions with 1 to 20 coefficients and uses the Akaike Information
Criterion (AIC) to evaluate the goodness of the fits. A model is considered a good fit, when the percentage of change from one model to the next (e.g. a model with
6 coefficients to a model with 7 coefficients) is \verb{< 5\%}. The first for models meeting this criterion are plotted as colored graphs and the AICs of these models
are visualized in a second plot for each curve. All first four coefficients per curve that fulfill the criterion are stored and in the end, a histogram of how
often which coefficients were good fits is plotted as well. The function returns the numerical value of the coefficient that fulfilled the criterion of a good fit
in most curves. If \code{print.to.pdf == TRUE}, then the plots are saved as PDFs in \code{path.plots}. Resulting data is saved in \code{path.data}.
}
\examples{
# Using the forceR::peaks.df.100.avg dataset:

# find smallest polynomial degree that best describes all curves
best.fit.poly <- find_best_fits(df = peaks.df.100.avg,
                                print.to.pdf = FALSE)

}
