% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot Peaks}
\usage{
plot_peaks(
  df.peaks,
  df.data,
  additional.msecs = 2000,
  path.plots,
  print.to.pdf = TRUE
)
}
\arguments{
\item{df.peaks}{df.peaks The resulting tibble of the function \code{find_peaks()}. See \code{?find_peaks} for more details.}

\item{df.data}{A data frame or tibble in the below format. The columns \code{t} (time), \code{force} and \code{measurement}
(measurement ID) must be present. This will usually be the same table that was used before in \code{find_peaks()}.}

\item{additional.msecs}{A numeric value indicating how many m.secs before and after the actual peak curve should be plotted. Default: \code{2000}}

\item{path.plots}{A string character defining where to save the plots.}

\item{print.to.pdf}{A logical value indicating if the peak curves should be saved as PDFs. Default: \code{TRUE}
Default: \code{TRUE}}
}
\value{
In addition, if \code{print.to.pdf == TURE},
Plots one graph per peak curve and, if \code{print.to.pdf == TURE}, saves all peak curves as one PDF at \code{path.plots}.
}
\description{
Plots the peaks identified by the function \code{find_peaks()}.
}
\details{
#' The input data frame or tibble \code{df.data} needs to contain the following columns:\tabular{ccc}{
   \strong{\code{t}} \tab \strong{\code{force}} \tab \strong{\code{measurement}} \cr
   \code{t.1} \tab \code{force.1} \tab \code{measurement.1} \cr
   \code{...} \tab \code{...} \tab \code{...} \cr
   \code{t.n} \tab \code{force.n} \tab \code{measurement.m} \cr
}
}
\examples{
# Using the forceR::peaks.df and forceR::df.all.200.tax datasets:

# plot peaks
plot_peaks(df.peaks = peaks.df,
           df.data = df.all.200.tax,
           additional.msecs = 20,
           print.to.pdf = FALSE)
}
